fix apd blank screen
