--- web/src/components/ApdApplication.js
@@ -1,36 +1,0 @@
-import React from 'react';
-
-// Wired-up components
-import StateProfile from '../components/ApdStateProfile';
-import AssurancesAndCompliance from '../containers/AssurancesAndCompliance';
-import ProposedBudget from './ProposedBudget';
-import Activities from '../containers/Activities';
-import ApdSummary from '../containers/ApdSummary';
-import CertifyAndSubmit from '../containers/CertifyAndSubmit';
-import ExecutiveSummary from '../containers/ExecutiveSummary';
-import PreviousActivities from '../containers/PreviousActivities';
-import ScheduleSummary from '../containers/ScheduleSummary';
-import Sidebar from '../containers/Sidebar';
-import TopBtns from '../containers/TopBtns';
-
-const PLACE = { id: 'tx', name: 'Texas' };
-
-const ApdApplication = () => (
-  <div className="site-body">
-    <Sidebar place={PLACE} />
-    <div className="site-main p2 sm-p4 md-px0">
-      <TopBtns />
-      <StateProfile />
-      <ApdSummary />
-      <PreviousActivities />
-      <Activities />
-      <ScheduleSummary />
-      <ProposedBudget />
-      <AssurancesAndCompliance />
-      <ExecutiveSummary />
-      <CertifyAndSubmit />
-    </div>
-  </div>
-);
-
-export default ApdApplication;
